requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography>=42.0.0
anthropic==0.94.1
httpx==0.28.1
httpcore==1.0.9
certifi==2026.2.25
pydantic-settings==2.1.0
python-dotenv==1.0.1
gunicorn==22.0.0
alembic==1.14.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
python-multipart>=0.0.6
python-docx
litellm>=1.40.0
# Optional: install Tesseract OCR binary separately for OCR image detection
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Ubuntu:  sudo apt install tesseract-ocr tesseract-ocr-vie
pytesseract>=0.3.10
PyMuPDF>=1.24.0
python-pptx>=1.0.0
openpyxl>=3.1.0

